+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - `soulstruct` updated to 2.4.0.
 - API for `Binder` changed for new `soulstruct` (`Firelink` compatibility).
+- `BothResHKXBHD.get_both_hkx()` split off into `BothResHKXBHD.get_both_hkx_allow_missing()`
 
 ## [1.2.2] - 2026-03-30